ability to change user role and protect the delet functionality

+++ convex/files.ts
@@ -2,7 +2,6 @@ import { ConvexError, v } from 'convex/values'
 import { MutationCtx, QueryCtx, mutation, query } from './_generated/server'
 import { getUser } from './users'
 import { fileTypes } from './schema';
-import { waitForDebugger } from 'inspector';
 import { Id } from './_generated/dataModel';
 
 export const generateUploadUrl = mutation(async (ctx) => {
@@ -11,10 +10,16 @@ export const generateUploadUrl = mutation(async (ctx) => {
     return await ctx.storage.generateUploadUrl();
 });
 
-async function hasAccessToOrg(ctx: QueryCtx | MutationCtx, tokenIdentifier: string, orgId: string) {
-    const user = await getUser(ctx, tokenIdentifier)
-    const hasAccess = user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId)
-    return hasAccess
+async function hasAccessToOrg(ctx: QueryCtx | MutationCtx, orgId: string) {
+    const idenity = await ctx.auth.getUserIdentity()
+    if (!idenity) return null
+
+    const user = await getUser(ctx, idenity.tokenIdentifier)
+
+    const hasAccess = user.orgIds.some(item => item.orgId === orgId) || user.tokenIdentifier.includes(orgId)
+    if (!hasAccess) return null
+
+    return { user }
 }
 
 export const createFile = mutation({
@@ -25,11 +30,7 @@ export const createFile = mutation({
         type: fileTypes
     },
     async handler(ctx, args) {
-        const idenity = await ctx.auth.getUserIdentity()
-
-        if (!idenity) throw new ConvexError("You must be logged in to upload/create a file")
-
-        const hasAccess = await hasAccessToOrg(ctx, idenity.tokenIdentifier, args.orgId)
+        const hasAccess = await hasAccessToOrg(ctx, args.orgId)
 
         if (!hasAccess) throw new ConvexError('you do not have access to this org')
 
@@ -49,12 +50,10 @@ export const getFiles = query({
         favorites: v.optional(v.boolean()),
     },
     async handler(ctx, args) {
-        const idenity = await ctx.auth.getUserIdentity()
-
-        if (!idenity) return []
-
-        const hasAccess = await hasAccessToOrg(ctx, idenity.tokenIdentifier, args.orgId)
+        const hasAccess = await hasAccessToOrg(ctx, args.orgId)
         if (!hasAccess) return []
+
+        const { user } = hasAccess
 
         const allFiles = await ctx.db
             .query("files")
@@ -62,15 +61,7 @@ export const getFiles = query({
             .collect()
 
         const query = args.query
-
         const filesSelected = query ? allFiles.filter(file => file.name.toLowerCase().includes(query.toLowerCase())) : allFiles
-
-        const user = await ctx.db
-            .query("users")
-            .withIndex("by_tokenIdentifier", q => q.eq("tokenIdentifier", idenity.tokenIdentifier))
-            .first()
-
-        if (!user) return filesSelected
 
         if (args.favorites) {
             const favorites = await ctx.db
@@ -80,17 +71,14 @@ export const getFiles = query({
                     .eq("orgId", args.orgId)
                 ).collect()
 
-
-                // if it's very slow I must refactor these type of filter systems 
+            // if it's very slow I must refactor these type of filter systems 
             return filesSelected.filter((file) =>
                 favorites.some(
                     (favorite) => favorite.fileId === file._id
                 )
             )
         }
-
         return filesSelected
-
     }
 })
 
@@ -100,6 +88,10 @@ export const deleteFile = mutation({
         const access = await hasAccessToFile(ctx, args.fileId)
         if (!access) throw new ConvexError("no access to file")
 
+        const isAdmin = access.user.orgIds.find(org => org.orgId === access.file.orgId)?.role === "admin"
+        
+        if(!isAdmin) throw new ConvexError("you have no admin access to delete")
+
         await ctx.db.delete(args.fileId)
     }
 })
@@ -108,9 +100,7 @@ export const toogleFavorite = mutation({
     args: { fileId: v.id('files') },
     async handler(ctx, args) {
         const access = await hasAccessToFile(ctx, args.fileId)
-
         if (!access) throw new ConvexError("no access to file")
-
         const { user, file } = access
 
         const favorite = await ctx.db
@@ -135,25 +125,35 @@ export const toogleFavorite = mutation({
     }
 })
 
+export const getAllFavorites = query({
+    args: { orgId: v.string() },
+    async handler(ctx, args) {
+        const hasAccess = await hasAccessToOrg(ctx, args.orgId)
+        if (!hasAccess) return null
+
+        const { user } = hasAccess
+
+        return await ctx.db
+            .query("favorites")
+            .withIndex("by_userId_orgId_fileId",
+                q => q
+                    .eq("userId", user._id)
+                    .eq("orgId", args.orgId)
+            )
+            .collect()
+
+
+    }
+})
+
 async function hasAccessToFile(ctx: QueryCtx | MutationCtx, fileId: Id<"files">) {
-    const idenity = await ctx.auth.getUserIdentity()
-
-    if (!idenity) return null
-
     const file = await ctx.db.get(fileId)
-
     if (!file) return null
 
-    const hasAccess = await hasAccessToOrg(ctx, idenity.tokenIdentifier, file.orgId)
-
+    const hasAccess = await hasAccessToOrg(ctx, file.orgId)
     if (!hasAccess) return null
 
-    const user = await ctx.db
-        .query("users")
-        .withIndex("by_tokenIdentifier", q => q.eq("tokenIdentifier", idenity.tokenIdentifier))
-        .first()
-
-    if (!user) return null
+    const { user } = hasAccess
 
     return { user, file }
 }
